+++ EFO/src/main/java/com/efo/interfaces/IPaymentsBilled.java
@@ -4,8 +4,8 @@ import com.efo.entity.PaymentsBilled;
 
 public interface IPaymentsBilled {
 	public void create(PaymentsBilled payments);
-	public PaymentsBilled retreive(int id);
+	public PaymentsBilled retreive(Long id);
 	public void update(PaymentsBilled payments);
-	public void delete(int id);
+	public void delete(Long id);
 
 }
