+++ EFO/src/main/java/com/efo/interfaces/ISalesItem.java
@@ -7,7 +7,7 @@ import com.efo.entity.SalesItem;
 public interface ISalesItem {
 	
 	public void create(SalesItem salesItem);
-	public SalesItem retrieve(int item_id);
+	public SalesItem retrieve(Long item_id);
 	public List<SalesItem> retrieveRawList();
 	public void update(SalesItem salesItem);
 	public void delete(SalesItem salesItem);
